Backend/requirements.txt

fastapi
uvicorn
python-multipart
nltk
spacy
textstat
transformers
sentence-transformers
torch
pydantic
python-jose
sqlalchemy
dotenv
psycopg2
passlib
PyJWT
emoji
bcrypt
openai